+++ packages/experimental/src/components/TagSet/TagSet.js
@@ -25,6 +25,7 @@ export const TagSet = ({
   className,
   maxVisibleTags,
   rightAlign,
+  overflowAlign,
   overflowDirection,
   showAllModalHeading,
   showAllSearchLabel,
@@ -72,13 +73,7 @@ export const TagSet = ({
         );
       } else {
         if (newOverflowTags.length < 10) {
-          newOverflowTags.push(
-            <span
-              key={`overflow-tag-${child.key}`}
-              className={`${blockClass}--overflow-tag`}>
-              {React.cloneElement(child)}
-            </span>
-          );
+          newOverflowTags.push(React.cloneElement(child));
         }
       }
     });
@@ -91,7 +86,9 @@ export const TagSet = ({
     if (showAllModalOpen) {
       const newFilteredAllTags = [];
       children.forEach((child) => {
-        const dataSearch = child.props['data-search']?.toLocaleLowerCase();
+        const dataSearch = (
+          child.props['data-search'] || ''
+        ).toLocaleLowerCase();
         const contentsAsString = child.props.children
           .toString()
           .toLocaleLowerCase();
@@ -179,9 +176,11 @@ export const TagSet = ({
   };
 
   const handleClickOutsideCheck = (ev) => {
-    const tooltipEl = overflowTagContent.current?.parentElement?.parentElement;
+    const tooltipEl =
+      overflowTagContent.current &&
+      overflowTagContent.current.parentElement.parentElement;
     if (
-      tooltipEl !== undefined &&
+      tooltipEl &&
       (tooltipEl === ev.target || tooltipEl.contains(ev.target))
     ) {
       // inside click
@@ -227,6 +226,8 @@ export const TagSet = ({
             })}
             onFocus={showTip}>
             <Tooltip
+              align={overflowAlign}
+              className={`${blockClass}--tooltip`}
               direction={overflowDirection}
               open={tipOpen}
               triggerText={<Tag>+{children.length - displayedTags.length}</Tag>}
@@ -235,7 +236,15 @@ export const TagSet = ({
               <div
                 ref={overflowTagContent}
                 className={`${blockClass}--overflow-content`}>
-                {overflowTags}
+                <ul className={`${blockClass}--overflow-tag-list`}>
+                  {overflowTags.map((tag, index) => (
+                    <li
+                      className={`${blockClass}--overflow-tag-item`}
+                      key={`overflow-tag--${index}`}>
+                      {tag.props.children}
+                    </li>
+                  ))}
+                </ul>
                 {overflowTags.length >= 10 && (
                   <Link
                     className={`${blockClass}--show-all-tags-link`}
@@ -285,6 +294,10 @@ TagSet.propTypes = {
    */
   maxVisibleTags: PropTypes.number,
   /**
+   * overflowAlign from the standard tooltip
+   */
+  overflowAlign: PropTypes.oneOf(['start', 'center', 'end']),
+  /**
    * overflowDirection from the standard tooltip
    */
   overflowDirection: PropTypes.oneOf(['top', 'right', 'bottom', 'left']),
@@ -311,6 +324,7 @@ TagSet.propTypes = {
 };
 
 TagSet.defaultProps = {
+  overflowAlign: 'center',
   overflowDirection: 'bottom',
   showAllModalHeading: 'All tags',
   showAllSearchLabel: 'Search all tags',