+++ packages/experimental/src/components/index.js
@@ -6,11 +6,12 @@
 //
 
 export { AboutModal } from './AboutModal';
+export { BreadcrumbWithOverflow } from './BreadcrumbWithOverflow';
 export { EmptyState } from './EmptyState';
 export { ExampleComponent } from './ExampleComponent';
 export { ImportModal } from './ImportModal';
 export { ModifiedTabs } from './ModifiedTabs';
 export { PageHeader, ActionBarItem, PageActionItem } from './PageHeader';
-export { RemovalModal } from './RemovalModal';
+export { RemoveDeleteModal } from './RemoveDeleteModal';
 export { TagSet } from './TagSet';
 export { Tearsheet, TearsheetNarrow } from './Tearsheet';