+++ packages/experimental/src/global/js/utils/wait.js
@@ -1,0 +1,7 @@
+// utility function for an async timeout
+
+function wait(ms) {
+  return new Promise((resolve) => setTimeout(resolve, ms));
+}
+
+export default wait;