+++ packages/jest-config/CHANGELOG.md
@@ -1,5 +1,48 @@
 # Change Log
 
+All notable changes to this project will be documented in this file.
+See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [0.3.0](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/jest-config@0.2.2...jest-config@0.3.0) (2020-12-16)
+
+
+### Features
+
+* Page header tag overflow update ([#264](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/264)) ([ee22520](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/ee225206e291fa3bca990cb2ccabfece8930fc88))
+
+
+
+
+
+## [0.2.2](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/jest-config@0.2.1...jest-config@0.2.2) (2020-12-08)
+
+**Note:** Version bump only for package jest-config
+
+
+
+
+
+## [0.2.1](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/jest-config@0.2.0...jest-config@0.2.1) (2020-11-30)
+
+**Note:** Version bump only for package jest-config
+
+
+
+
+
+# [0.2.0](https://github.com/carbon-design-system/ibm-cloud-cognitive/tree/master/packages/jest-config/compare/jest-config@0.1.1...jest-config@0.2.0) (2020-11-26)
+
+
+### Features
+
+* add cdai apikey component ([#223](https://github.com/carbon-design-system/ibm-cloud-cognitive/tree/master/packages/jest-config/issues/223)) ([4c09f15](https://github.com/carbon-design-system/ibm-cloud-cognitive/tree/master/packages/jest-config/commit/4c09f15c3c62c3965d98c91b9695fa7a1cba8f0b))
+
+
+
+
+
+# Change Log
+
 All notable changes to this project will be documented in this file. See
 [Conventional Commits](https://conventionalcommits.org) for commit guidelines.