+++ packages/storybook-addon-carbon-theme/CHANGELOG.md
@@ -1,5 +1,40 @@
 # Change Log
 
+All notable changes to this project will be documented in this file.
+See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [0.3.0](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/storybook-addon-theme@0.2.3...@carbon/storybook-addon-theme@0.3.0) (2020-12-16)
+
+
+### Features
+
+* Page header tag overflow update ([#264](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/264)) ([ee22520](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/ee225206e291fa3bca990cb2ccabfece8930fc88))
+
+
+
+
+
+## [0.2.3](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/storybook-addon-theme@0.2.2...@carbon/storybook-addon-theme@0.2.3) (2020-12-10)
+
+
+### Bug Fixes
+
+* stylelint plugin readme ([#270](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/270)) ([0aa1985](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/0aa19854c442c8d09ee3e9e482a019955c63282e))
+
+
+
+
+
+## [0.2.2](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/storybook-addon-theme@0.2.1...@carbon/storybook-addon-theme@0.2.2) (2020-11-30)
+
+**Note:** Version bump only for package @carbon/storybook-addon-theme
+
+
+
+
+
+# Change Log
+
 All notable changes to this project will be documented in this file. See
 [Conventional Commits](https://conventionalcommits.org) for commit guidelines.